+++ index.d.ts
@@ -481,7 +481,8 @@ declare module '@amaas/amaas-core-sdk-js' {
     businessDate: string
     clientId: string
     fxPnl: any
-    message: string
+    errorMessage: string
+    additional: any
     period: any
     pnlStatus: string
     currency: string
@@ -505,7 +506,8 @@ declare module '@amaas/amaas-core-sdk-js' {
     businessDate: string
     clientId: string
     fxPnl: any
-    message: string
+    errorMessage: string
+    additional: any
     period: any
     pnlStatus: string
     pnlTimestamp: string
@@ -538,7 +540,9 @@ declare module '@amaas/amaas-core-sdk-js' {
       fx: string
       asset: string
     }
-    fxRates: any
+    fxRates: {
+      [rateName: string]: any
+    }
   }
 
   // Transactions
@@ -1429,15 +1433,20 @@ declare module '@amaas/amaas-core-sdk-js' {
       ): Promise<IMTMResult[]> | void
     }
     function config({
-      stage,
       credentialsPath,
-      apiVersion,
-      token
+      stage, apiURL, apiVersion,
+      cognitoPoolId, cognitoClientId,
+      token, username, password,
     }: {
-      stage?: string
       credentialsPath?: string
+      stage?: string
+      apiURL?: string
       apiVersion?: string
+      cognitoPoolId?: string
+      cognitoClientId?: string
       token?: string
+      username?: string
+      password?: string
     }): void
     namespace PositionPNL {
       function retrieve(
@@ -2573,6 +2582,9 @@ declare module '@amaas/amaas-core-sdk-js' {
         asset: any
         fx: any
       }
+      fxRates: {
+        [rateName: string]: any
+      }
       constructor(props: IAggregatePNL)
     }
 
@@ -2605,7 +2617,8 @@ declare module '@amaas/amaas-core-sdk-js' {
       clientId: string
       fxPnl: any
       currency: string
-      message: string
+      errorMessage: string
+      additional: any
       period: any
       pnlStatus: string
       pnlTimestamp: string
@@ -2629,7 +2642,7 @@ declare module '@amaas/amaas-core-sdk-js' {
       businessDate: string
       clientId: string
       fxPnl: any
-      message: string
+      errorMessage: string
       period: any
       pnlStatus: string
       pnlTimestamp: string
@@ -2644,6 +2657,7 @@ declare module '@amaas/amaas-core-sdk-js' {
       version?: string
       transactionId: string
       currency: string
+      additional: any
       constructor(props: ITransactionPNL)
     }
   }