+++ src/transactions/AggregatePNL/aggregatePNL.js
@@ -1,4 +1,5 @@
 import Decimal from 'decimal.js'
+import mapValues from 'lodash/mapValues'
 
 import { AMaaSModel } from '../../core'
 
@@ -8,10 +9,19 @@ import { AMaaSModel } from '../../core'
  * @extends module:core.AMaaSModel
  */
 class AggregatePNL extends AMaaSModel {
+  /**
+   * Construct a new AggregatePNL object
+   * @param {object} params - AggregatePNL creation options
+   * @param {object} params.YTD - YTD PnL containing `fx`, `asset` and `total` PnL values`
+   * @param {object} params.MTD - MTD PnL containing `fx`, `asset` and `total` PnL values`
+   * @param {object} params.DTD - DTD PnL containing `fx`, `asset` and `total` PnL values`
+   * @param {object} params.fxRates - FX rates. This is an object with the ccy pair as key and rate as value
+   */
   constructor({
     YTD,
     MTD,
     DTD,
+    fxRates,
     createdBy,
     updatedBy,
     createdTime,
@@ -61,12 +71,25 @@ class AggregatePNL extends AMaaSModel {
           this._DTD = { total, fx, asset }
         },
         enumerable: true
+      },
+      _fxRates: { writable: true, enumerable: false },
+      fxRates: {
+        get: () => this._fxRates,
+        set: (newFxRates = {}) => {
+          let castFxRates = mapValues(
+            newFxRates,
+            rate => new Decimal(rate || 0)
+          )
+          this._fxRates = castFxRates
+        },
+        enumerable: true
       }
     })
 
     this.YTD = YTD
     this.MTD = MTD
     this.DTD = DTD
+    this.fxRates = fxRates
   }
 }