+++ src/utils/aggregatepnl/aggregatepnl.test.js
@@ -1,5 +1,8 @@
+import Decimal from 'decimal.js'
+
 import * as api from '../../exports/api'
 import * as network from '../network'
+import AggregatePNL from '../../transactions/AggregatePNL/aggregatePNL'
 
 network.retrieveData = jest.fn()
 
@@ -26,6 +29,29 @@ describe('retrieve', () => {
           AMId: 88,
           query: { key: 'value' }
         })
+        done()
+      }
+    )
+  })
+  it('returns a class', done => {
+    network.retrieveData.mockImplementation(() =>
+      Promise.resolve({
+        fxRates: { HKDUSD: '1.23' },
+        aggregatePnl: {
+          DTD: {
+            total: 8,
+            asset: 3,
+            fx: 5
+          }
+        }
+      })
+    )
+    api.AggregatePNL.retrieve(
+      { AMId: 8, quuery: { key: 'value' } },
+      (error, result) => {
+        expect(result).toBeInstanceOf(AggregatePNL)
+        expect(result.fxRates.HKDUSD).toEqual(new Decimal('1.23'))
+        expect(result.DTD.total).toEqual(new Decimal(8))
         done()
       }
     )